Fix implicit license plans
Licenses without a plan specified need to act as though they are EES. This is
currently implemented for licenses that don't include a plan in their data at
all, but if `plan: nil` or `plan: ""` is set, the customer loses access to a
number of features with license checks added in 9.4. This commit changes the
license check to use the starter plan in those cases too.