Fix diff commenting results just after changing view
When you change the diff view (inline / side-by-side), we set a cookie based on
that new view. When you add a comment, we choose the style to use in the
response based on that cookie.

However, when you have just changed diff style, the request cookie will contain
the old value, so we should use the view param instead.