Don't bother going through an entire Banzai pipeline for empty text
This bails out of `cacheless_render` immediately unless the provided
text is present, since there's no point.

This is a slight improvement in our test performance. Across the
creation of 1,000 `Namespace` records, which caches the `description`
field and which is blank by default in its factory, this saves about
four seconds, which... sure. Why not.