Fix for spec/requests/api/boards_spec.rb
In gitlab-org/gitlab-ce!13283 extra properties were added to
BasicProjectDetails. This entity is used by the Board entity and the
API boards spec verifies the JSON schema for the
`/projects/:id/boards` endpoint.

So add the extra properties to the schema file.

A better solution would have been to extract the project into a
separate file. But that is outside the current scope.

This also correctly fixes the conflict introduced in 3c8b2ae168, and
resolved in 55a5f2ab49.