Precompile resolve button Vue template
With a pre-compiled template we get a massive performance boost that can improve the initial rendering of the notes. Currently Vue has to go through each button & then compile the template & then finally render this compiled template into the DOM. With this we skip the compiling stage by pre-compiling the Vue templates through .vue files.

The number of watchers in this component been reduced to help with the initial initializing of the Vue app