Migrate down before each migration unit test
`migrate!` helper triggers a migration that is under the test, so we
need to revert it before each subsequent example. This means that we
need to place it in `before(:each, :migration)` hook.

We still want to migrate everything up *after* each *context*, so in
otherwords we need to place migrate up helper in `before(:context,
:migration)`.