Use pagination without COUNT for issues/MRs
This changes the pagination of issues and merge request index pages so
they don't use page numbers, instead only displaying a "Next" and
"Previous" button. This removes the need for a COUNT(*) query which on
GitLab.com can easily take up to 30 milliseconds.

A downside of this approach is that if you specify an offset that no
longer leads to data (e.g. because issues were closed) you won't be
redirected to the last valid page. This can't be done since this
functionality relies on us knowing the total number of pages, which in
turn requires a COUNT(*) query.