Remove skipped examples in filtered issues feature spec
We perform a ton of expensive setup in this spec, via both `before` and
`let!`. When we use `skip` inside an `it` block, rather than skipping
the block entirely via `xit`, we still perform this expensive setup,
only to do absolutely nothing with it, resulting in wasted time.
Minutes, in this case.

Further, these specs have been sitting in our test suite for about seven
months now, so the likelihood of them being addressed soon seems low.