Remove redundant specs from filter_issues_spec
While testing the various search tokens, we would add the token and a
search term, then verify the correct issues were returned. Then we'd add
a new token, search again, and verify again. Then another, and another.

This was entirely redundant. If the search with higher specificity
returned the same results as the search with lower specificity, we can
be sure the lower specificity searches also worked without actually
testing them.

Because this entire spec performs very extensive setup, it's important
to limit the number of times we perform that setup.