Clean up the still-way-too-complex filter_issues_spec setup
Removes unnecessary `let!` usage since we explicitly access these in the
`before.

Re-uses `user2` as the author for issues we otherwise don't care about
the author for, to avoid creating a new user in each one.

Makes use of the `labeled_issue` factory to avoid assigning variables
just so we can add labels to an issue.