Do not show YAML frontmatter for doc pages under /help
We recently started adding YAML frontmatter in docs so that we can
show more information, but that only works for the docs portal at
docs.gitlab.com.

For example, we want to add a last_updated entry
https://gitlab.com/gitlab-org/gitlab-ce/issues/37677

Whereas this is useful for the docs portal, it looks ugly for docs
under /help.