Remove an unneeded index on ci_builds left in some databases
Migration 20170919211300_remove_temporary_ci_builds_index.rb created a
temporary partial index and tried to drop it at the end of the
migration. In some circumstances apparently it failed to drop the
index and it ended up in our schema.rb.

This accidentally failed to fail due to a bug in the regular
expression for partial indexes which caused the index creation in
schema.rb to be ignored. Now that that's fixed we could be
resurrecting this zombie index from the past in some but not all
databases.

Add a migration to drop this index if it's present to reconcile this
discrepancy.