Remove varchar length limits leftover in databases that were migrateid from MySQL in the past. This means altering any tables that have varchar(255) or varchar(510) to just be plain varchar. This brings these databases closer into sync with the bundled schema.rb.
This will not require rewriting any tables and will not require
scanning the entirety of any tables so it should be very
fast.

Nonetheless it does require an exclusive lock on each affected table
(one at a time, not all at once) so on a very busy database it could
cause a short period of lack of responsiveness on the database. It's
important as a result that this run at a time when no long-running
background migrations or batch queries are running on the database.