Do not clone the repo when running the review-docs jobs
Previously, we checked out the branch and ran the script from there.
However, when the MR got merged and the cleanup job was run it would
fail since it could not find the now deleted branch.

This commit rather downloads the script from GitLab and skips
downloading the artifacts/dependencies/cache to speed up the process.