[Rails 5] Fix the problem around default values of push_events and repository_update_events
The default_value_for gem does not work for rails 5 as expected. The initial fix was https://gitlab.com/gitlab-org/gitlab-ce/commit/4b4708e50dccfcba5cc65ead58853a54685b4e77 but after some time I decided to chage the approach.

The thing is that we have default values in a database that  satisfies us for web_hooks entities. But
for system hooks, we need different values. In fact, we don't need the default_value_for gem to set different default values.
We can merely set the different default values in API and views(forms) for system hooks. Because why not???
This particular commit implements the idea.