Fixes commit comments in side-by-side diff view
This was caused by the `notes` global class not existing when the
`file_comment_button` code is run. The notes class was used to check
if the diff is currently in parallel view or not. To get around this
I've added a check into the `file_comment_button` JS to check if the
view is currently parallel or not.

Closes #38117