Fix fork button being disabled for users who can fork to group
Previously the fork button was disabled for all users if they have
exceeded their project limit. This fixes that by changing the check
to see if the user can fork to a group instead of their own namespace.

This behaviour is already possible by visiting the new fork page
directly, so this just fixes the button being disabled.

Closes #38462