Show stale instead of merged if the branch is old
We consider the branch is old when the updated date was
3 months older than the default branch updated date.

By doing so, it would greatly improve the performance
in this page, because it's pretty costly to find if it's
merged or not when the branch is far behind.

We assume that when it's old, it's probably far behind.

The idea was brought up in:
https://gitlab.com/gitlab-org/gitlab-ce/issues/37429#note_42284280