don't filter group runners by project flag
the scope `Ci::Runner.belonging_to_group` does not filter out the
runners where the projects has `#group_runners_enabled` set to false
anymore.

it didn't show up in the runners UI anymore when group runners were
disabled. this was confusing.

the flag is only relevant when selecting appropriate runner for a build.