Fix wordings and always make is_merged accurate.
At first I want to tweak the wordings so that people would
be aware that it might be stale rather than not merged yet.
But that also involved translations...

After some second thought, we would only try to calculate this if:

* `can?(current_user, :push_code, @project)`
* `protected_branch?(@project, branch)`
* `can?(current_user, :delete_protected_branch, @project)`

And this might already reduce the calculation for quite a lot.
We could revisit the performance later, if this is still slow.
Better get it right than trying to be fast without knowing if
it's an issue.