QA group name, use hyphens rather than underscores
In doing some experimental testing of enabling the embedded mattermost
we found that the currently generated QA namespace was problematic in that
it used underscores when generating group names.  This is invalid for a
mattermost team name, underscores are illegal there.

Here we change to use hyphens as they are legal in team names.